services/resume_job_analyzer/requirements.txt
Flask~=3.1.2
flask-cors~=4.0.0

openai~=2.6.0
python-dotenv

PyPDF2~=3.0.1
python-docx
lxml

werkzeug~=3.1.3
gunicorn